+++ optd-cost-model/src/lib.rs
@@ -33,8 +33,8 @@ pub struct Cost(pub Vec<f64>);
 
 /// Estimated statistic calculated by the cost model.
 /// It is the estimated output row count of the targeted expression.
-#[derive(Eq, Ord, PartialEq, PartialOrd, Debug)]
-pub struct EstimatedStatistic(pub u64);
+#[derive(PartialEq, PartialOrd, Debug)]
+pub struct EstimatedStatistic(pub f64);
 
 pub type CostModelResult<T> = Result<T, CostModelError>;
